requirements.txt
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
filelock==3.20.0
Flask==3.1.2
fsspec==2025.10.0
google-api-core==2.28.1
google-auth==2.42.1
google-cloud-core==2.5.0
google-cloud-translate==3.22.0
googleapis-common-protos==1.71.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.0.1
regex==2025.10.23
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
setuptools==80.9.0
spotipy==2.25.1
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
